refactor(api): move test helpers to own file
